requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
yfinance>=0.2.33
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
statsmodels>=0.14.0
torch>=2.1.1
mlflow>=2.8.1
evidently>=0.4.7
APScheduler>=3.10.4
python-dotenv>=1.0.0
PyYAML>=6.0.1
